Fix Debian dependency issue (#154)
Fixes the dependency string of the Debian package for the Nightly PPA.
Thereby resolves #136.

There are now some containers available to verify this:
https://github.com/a-sit/ppa-nightly/tree/main/.github/testing

As expected, the install now works for Ubuntu 24.04 and newer, and the
upcoming Debian Trixie (13), but _not_ the current Debian Bookworm (12).
[Debian Tricksie](https://wiki.debian.org/DebianTrixie) is already in
hard freeze and will be released soon, so I won't fix the issue for
bookworm (Java version too old).